+++ test/spec/event.js
@@ -62,7 +62,7 @@ describe('event form attribute callback global', function() {
     });
 
     it('call data-bv-onerror', function() {
-        this.$email.val('email@domain');
+        this.$email.val('a@b@c@example.com');
         this.bv.validate();
         expect($('#msg').html()).toEqual('form eventForm is invalid');
     });
@@ -96,7 +96,7 @@ describe('event form attribute callback namespace', function() {
     });
 
     it('call data-bv-onerror', function() {
-        this.$email.val('email@domain');
+        this.$email.val('just"not"right@example.com');
         this.bv.validate();
         expect($('#msg').html()).toEqual('TestSuite.Event.onFormInvalid() called, form eventForm is invalid');
     });
@@ -137,7 +137,7 @@ describe('event form trigger', function() {
     });
 
     it('trigger error.form.bv', function() {
-        this.$email.val('email@domain');
+        this.$email.val('this is"not\\allowed@example.com');
         this.bv.validate();
         expect($('#msg').html()).toEqual('form eventForm triggered error.form.bv event');
     });
@@ -178,7 +178,7 @@ describe('event form programmatically', function() {
     });
 
     it('call onError()', function() {
-        this.$email.val('email@domain');
+        this.$email.val('Abc.example.com');
         this.bv.validate();
         expect($('#msg').html()).toEqual('onError() called');
     });
@@ -230,7 +230,7 @@ describe('event field attribute callback global', function() {
     });
 
     it('call data-bv-onerror', function() {
-        this.$email.val('email@domain');
+        this.$email.val('A@b@c@example.com');
         this.bv.validate();
         expect($('#msg').html()).toEqual('email is invalid');
         expect($('#status').html()).toEqual(this.bv.STATUS_INVALID);
@@ -267,7 +267,7 @@ describe('event field attribute callback namespace', function() {
     });
 
     it('call data-bv-onerror', function() {
-        this.$email.val('email@domain');
+        this.$email.val('a"b(c)d,e:f;gi[j\\k]l@example.com');
         this.bv.validate();
         expect($('#msg').html()).toEqual('TestSuite.Event.onEmailInvalid() called, email is invalid');
         expect($('#status').html()).toEqual('TestSuite.Event.onEmailStatus() called; status = ' + this.bv.STATUS_INVALID);
@@ -309,7 +309,7 @@ describe('event field trigger', function() {
     });
 
     it('trigger error.field.bv', function() {
-        this.$email.val('email@domain');
+        this.$email.val('just"not"right@example.com');
         this.bv.validate();
         expect($('#msg').html()).toEqual('triggered error.field.bv on email');
     });
@@ -354,7 +354,7 @@ describe('event field programmatically', function() {
     });
 
     it('call onError()', function() {
-        this.$email.val('email@domain');
+        this.$email.val('this is"not\\allowed@example.com');
         this.bv.validate();
         expect($('#msg').html()).toEqual('onError() called');
     });
@@ -411,13 +411,13 @@ describe('event form trigger with default events', function() {
     });
 
     it('does not trigger bv.form.error', function() {
-        this.$email.val('email@domain');
+        this.$email.val('A@b@c@example.com');
         this.bv.validate();
         expect($('#msg').html()).not.toEqual('form eventForm1 triggered bv.form.error event');
     });
 
     it('triggers error.form.bv', function() {
-        this.$email.val('email@domain');
+        this.$email.val('A@b@c@example.com');
         this.bv.validate();
         expect($('#msg').html()).toEqual('form eventForm1 triggered error.form.bv event');
     });
@@ -470,13 +470,13 @@ describe('event field trigger with default events', function() {
     });
 
     it('does not trigger error.field.bv', function() {
-        this.$email.val('email@domain');
+        this.$email.val('just"not"right@example.com');
         this.bv.validate();
         expect($('#msg').html()).toEqual('triggered error.field.bv on email');
     });
 
     it('triggers bv.field.error', function() {
-        this.$email.val('email@domain');
+        this.$email.val('just"not"right@example.com');
         this.bv.validate();
         expect($('#msg').html()).not.toEqual('triggered bv.field.error on email');
     });
@@ -547,7 +547,7 @@ describe('event form trigger with events changed', function() {
     it('triggers bv.form.error', function() {
         spyOn(window, 'onerror');
 
-        this.$email.val('email@domain');
+        this.$email.val('this is"not\\allowed@example.com');
         this.bv.validate();
         expect($('#msg').html()).toEqual('form eventForm2 triggered bv.form.error event');
 
@@ -618,7 +618,7 @@ describe('event field trigger with events changed', function() {
     });
 
     it('does not trigger error.field.bv', function() {
-        this.$email.val('email@domain');
+        this.$email.val('Abc.example.com');
         this.bv.validate();
         expect($('#msg').html()).not.toEqual('triggered error.field.bv on email');
     });
@@ -626,7 +626,7 @@ describe('event field trigger with events changed', function() {
     it('triggers bv.field.error', function() {
         spyOn(window, 'onerror');
 
-        this.$email.val('email@domain');
+        this.$email.val('Abc.example.com');
         this.bv.validate();
         expect($('#msg').html()).toEqual('triggered bv.field.error on email');
 
@@ -674,7 +674,7 @@ describe('event validator declarative', function() {
     });
 
     it('trigger data-bv-emailaddress-onerror', function() {
-        this.$email.val('email@domain');
+        this.$email.val('A@b@c@example.com');
         this.bv.validate();
         expect($('#msg').html()).toEqual('emailAddress validator did not pass');
     });
@@ -724,7 +724,7 @@ describe('event validator programmatically', function() {
     });
 
     it('call onError()', function() {
-        this.$email.val('email@domain');
+        this.$email.val('A@b@c@example.com');
         this.bv.validate();
         expect($('#msg').html()).toEqual('emailAddress validator: onError() called');
     });